Big updates
Ok. So... I'm on the fence about moving all this to google sheets or going the fancy python selenium route. I'm unable to get a webhook or api to grab active picks. Basically, I want to not do much thinking in this if possible, and if I were to manually check a box in my spreadsheet every second, that would be ass. So... kind of feels like python might be the way to go? This would require a few things. Adding an interface for me, probably text, to see a list of the top remaining players, the top remaining rookie/lotto/handcuff players and then I make a selection. either way, I'm needing to make a decision. I can either trust my word search abilities and grab player names that way, or find a way to automate it.